+++ app/controllers/application_controller.rb
@@ -1,4 +1,6 @@
 class ApplicationController < ActionController::Base
+  include Authentication
+
   layout "inertia"
 
   inertia_share do